// src/api/generated/core.ts
export interface components {
    schemas: {
// ...
        CabinetParityRolloutClaimDto: {
            audiences: string[];
            grants: string[];
            subjectId: string;
            /** Format: int32 */
            version: number;
        };
        CabinetParityRolloutEnvelopeDto: {
            claim: components["schemas"]["CabinetParityRolloutClaimDto"];
            configuration: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}